fix: - moved LoadingIndicator component to new utility folder
